requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.14.0
google-generativeai==0.3.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.0
psycopg2-binary==2.9.9
django-filter==23.4
django-cors-headers==4.3.1
gunicorn==21.2.0
